[Kashif]Added: BeanUtils and Swagger config

+++ src/main/java/com/cg/addressbook/service/AddressBookService.java
@@ -1,8 +1,8 @@
 package com.cg.addressbook.service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,7 +11,10 @@ import com.cg.addressbook.exceptions.AddressBookException;
 import com.cg.addressbook.model.AddressBookData;
 import com.cg.addressbook.repository.AddressBookRepository;
 
+import lombok.extern.slf4j.Slf4j;
+
 @Service
+@Slf4j
 public class AddressBookService implements IAddressBookService {
 
 	@Autowired
@@ -23,40 +26,30 @@ public class AddressBookService implements IAddressBookService {
 	}
 
 	@Override
-	public AddressBookData getAddressBookContactDataById(int id) {
-		try {
-			AddressBookData contactData = addressBookRepository.findById(id).get();
-			return contactData;
-		} catch (NoSuchElementException exception) {
-			throw new AddressBookException("Address Book Contact Not Found");
-		}
+	public AddressBookData getAddressBookContactDataById(Long id) {		
+			AddressBookData contactData = addressBookRepository.findById(id)
+											.orElseThrow(() -> new AddressBookException("Address Book Contact Not Found"));
+			return contactData;	
 	}
 
 	@Override
-	public AddressBookData createAddressBookContactData(AddressBookDTO addressBookDTO) {
+	public AddressBookData createAddressBookContactData(AddressBookDTO addressBookDTO) {		
 		AddressBookData contactData = null;
 		contactData = new AddressBookData(addressBookDTO);
+		log.debug("AddressBookData"+contactData.toString());
 		return addressBookRepository.save(contactData);
 	}
 
 	@Override
-	public AddressBookData updateAddressBookContactData(int id, AddressBookDTO addressBookDTO) {
+	public AddressBookData updateAddressBookContactData(Long id, AddressBookDTO addressBookDTO) {
 		AddressBookData contactData = this.getAddressBookContactDataById(id);
-		if (contactData != null) {
-			deleteAddressBookContactData(id);
-			contactData.setFullName(addressBookDTO.fullName);
-			contactData.setAddress(addressBookDTO.address);
-			contactData.setCity(addressBookDTO.city);
-			contactData.setState(addressBookDTO.state);
-			contactData.setZip(addressBookDTO.zip);
-			contactData.setPhoneNumber(addressBookDTO.phoneNumber);
-			contactData = addressBookRepository.save(contactData);
-		}
-		return contactData;
+		BeanUtils.copyProperties(addressBookDTO, contactData);
+		return addressBookRepository.save(contactData);
 	}
 
 	@Override
-	public void deleteAddressBookContactData(int id) {
-		addressBookRepository.deleteById(id);
+	public void deleteAddressBookContactData(Long id) {
+		if(this.getAddressBookContactDataById(id) != null)
+			addressBookRepository.deleteById(id);
 	}
 }
